refactor(logs): rename file
