+++ client/app/bundles/course/assessment/submission/pages/LogsIndex/index.tsx
@@ -1,5 +1,3 @@
-import { LogInfo } from 'types/course/assessment/submission/logs';
-
 import LoadingIndicator from 'lib/components/core/LoadingIndicator';
 import PageHeader from 'lib/components/navigation/PageHeader';
 import Preload from 'lib/components/wrappers/Preload';
@@ -14,29 +12,25 @@ import LogsHead from './LogsHead';
 import translations from './translations';
 
 const LogsIndex = (): JSX.Element => {
-  const fetchLogsData = (): Promise<LogInfo> => fetchLogs();
-
   const { t } = useTranslation();
 
   return (
-    <Preload render={<LoadingIndicator />} while={fetchLogsData}>
-      {(data): JSX.Element => {
-        return (
-          <>
-            <PageHeader
-              returnLink={getAssessmentSubmissionURL(
-                getCourseId(),
-                getAssessmentId(),
-              )}
-              title={t(translations.accessLogs)}
-            />
-            <div id="submission-log">
-              <LogsHead with={data.info} />
-              <LogsContent with={data.logs} />
-            </div>
-          </>
-        );
-      }}
+    <Preload render={<LoadingIndicator />} while={fetchLogs}>
+      {(data): JSX.Element => (
+        <>
+          <PageHeader
+            returnLink={getAssessmentSubmissionURL(
+              getCourseId(),
+              getAssessmentId(),
+            )}
+            title={t(translations.accessLogs)}
+          />
+          <div id="submission-log">
+            <LogsHead with={data.info} />
+            <LogsContent with={data.logs} />
+          </div>
+        </>
+      )}
     </Preload>
   );
 };
